requirements.txt
# ===============================
# Web Server
# ===============================
fastapi==0.127.0
uvicorn==0.40.0

# ===============================
# LangChain (Classic / Stable)
# ===============================
langchain==1.2.0
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langchain-chroma==1.1.0

# ===============================
# Vector Store
# ===============================
chromadb==1.4.0

# ===============================
# Re-Ranking
# ===============================
FlashRank==0.2.10

# ===============================
# OpenAI
# ===============================
openai==2.14.0
tiktoken==0.12.0

# ===============================
# PDF Processing
# ===============================
PyMuPDF==1.24.0  # PDF 텍스트 추출 및 렌더링 (fitz)

# ===============================
# OCR (Optical Character Recognition)
# ===============================
pytesseract==0.3.13  # Tesseract OCR Python 래퍼
Pillow==11.0.0  # 이미지 처리 (PIL)

# ===============================
# Document Loaders
# ===============================
langchain-community==1.0.0  # LangChain 커뮤니티 로더들 (TextLoader, PyPDFLoader, Docx2txtLoader, BSHTMLLoader)

# ===============================
# Data / Utils
# ===============================
pydantic==2.12.5
numpy==2.4.0
python-dotenv==1.2.1
